backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "blinker==1.8.2",
    "click==8.1.7",
    "flask==3.0.3",
    "flask-cors==6.0.0",
    "flask-marshmallow>=1.3.0",
    "flask-migrate==4.0.7",
    "flask-restful==0.3.10",
    "flask-restful-swagger-3==0.5.1",
    "flask-sqlalchemy==3.1.1",
    "greenlet==3.1.1",
    "gunicorn==23.0.0",
    "ipython>=9.5.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "marshmallow==3.23.1",
    "marshmallow-sqlalchemy>=1.4.2",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.1",
    "sqlalchemy==2.0.36",
    "typing-extensions==4.12.2",
    "werkzeug==3.0.6",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "tox==4.23.2",
]
